Modernize packaging: setup.py + setup.cfg --> pyproject.toml
Also moved the plugin from `__init__` to `plugin` for consistency with
other (modernized) first-party plugins and our plugin template.
