--- setup.cfg
@@ -1,26 +1,0 @@
-[metadata]
-name = sopel-py
-version = 1.0.1
-description = Sopel Python Eval Plugin
-author = dgw
-author_email = dgw@technobabbl.es
-url = https://github.com/sopel-irc/sopel-py
-license = Eiffel Forum License, version 2
-classifiers =
-    Intended Audience :: Developers
-    Intended Audience :: System Administrators
-    License :: Eiffel Forum License (EFL)
-    License :: OSI Approved :: Eiffel Forum License
-    Topic :: Communications :: Chat :: Internet Relay Chat
-
-[options]
-packages = find:
-zip_safe = false
-include_package_data = true
-install_requires =
-    sopel>=7.1,<9
-python_requires = >=3.7
-
-[options.entry_points]
-sopel.plugins =
-    py = sopel_py
